vanilla functionals added
